Extract episode creation
So it can be used in unit tests
