+++ Modules/Server/Sources/PocketCastsServer/Public/ServerPodcastManager.swift
@@ -258,43 +258,7 @@ public class ServerPodcastManager: NSObject {
 
         var episodes = [Episode]()
         for episodeJson in episodesJson {
-            let episode = Episode()
-            episode.addedDate = Date()
-            episode.podcast_id = podcast.id
-            episode.podcastUuid = podcast.uuid
-            episode.playingStatus = PlayingStatus.notPlayed.rawValue
-            episode.episodeStatus = DownloadStatus.notDownloaded.rawValue
-            if let uuid = episodeJson["uuid"] as? String {
-                episode.uuid = uuid
-            }
-            if let title = episodeJson["title"] as? String {
-                episode.title = title
-            }
-            if let url = episodeJson["url"] as? String {
-                episode.downloadUrl = url
-            }
-            if let fileType = episodeJson["file_type"] as? String {
-                episode.fileType = fileType
-            }
-            if let fileSize = episodeJson["file_size"] as? Int64 {
-                episode.sizeInBytes = fileSize
-            }
-            if let duration = episodeJson["duration"] as? Double {
-                episode.duration = duration
-            }
-            if let publishedStr = episodeJson["published"] as? String {
-                episode.publishedDate = isoFormatter.date(from: publishedStr)
-            }
-            if let number = episodeJson["number"] as? Int64 {
-                episode.episodeNumber = number
-            }
-            if let season = episodeJson["season"] as? Int64 {
-                episode.seasonNumber = season
-            }
-            if let type = episodeJson["type"] as? String {
-                episode.episodeType = type
-            }
-
+            let episode = Episode.from(episodeJson: episodeJson, podcastId: podcast.id, podcastUuid: podcast.uuid, isoFormatter: isoFormatter)
             episodes.append(episode)
         }
         DataManager.sharedManager.bulkSave(episodes: episodes)
